requirements.txt
aiogram
openai==0.27.0
pyTelegramBotAPI
python-dotenv